Split REST functionality into a tree-shakable module
This moves the following functionality from BaseClient into a new module
named Rest:

- methods that wrap REST endpoints (e.g. `stats`)
- the `request` method
- all functionality accessed via `BaseRest.channels` and `BaseRest.push`

This allows us to now construct a BaseRealtime instance that doesn’t
have REST functionality.

Note that the BaseRest class always includes the Rest module. This comes
after discussions in which we decided that it would be quite useless
without it.

Resolves #1374.
