+++ test/browser/modules.test.js
@@ -1,4 +1,4 @@
-import { BaseRest, BaseRealtime } from '../../build/modules/index.js';
+import { BaseRest, BaseRealtime, Rest } from '../../build/modules/index.js';
 
 describe('browser/modules', function () {
   this.timeout(10 * 1000);
@@ -12,11 +12,36 @@ describe('browser/modules', function () {
 
   describe('without any modules', () => {
     for (const clientClass of [BaseRest, BaseRealtime]) {
-      it(clientClass.name, async () => {
-        const client = new clientClass(ablyClientOptions());
+      describe(clientClass.name, () => {
+        it('can be constructed', async () => {
+          expect(() => new clientClass(ablyClientOptions(), {})).not.to.throw();
+        });
+      });
+    }
+  });
+
+  describe('Rest', () => {
+    describe('BaseRest without explicit Rest', () => {
+      it('offers REST functionality', async () => {
+        const client = new BaseRest(ablyClientOptions(), {});
         const time = await client.time();
         expect(time).to.be.a('number');
       });
-    }
+    });
+
+    describe('BaseRealtime with Rest', () => {
+      it('offers REST functionality', async () => {
+        const client = new BaseRealtime(ablyClientOptions(), { Rest });
+        const time = await client.time();
+        expect(time).to.be.a('number');
+      });
+    });
+
+    describe('BaseRealtime without Rest', () => {
+      it('throws an error when attempting to use REST functionality', async () => {
+        const client = new BaseRealtime(ablyClientOptions(), {});
+        expect(() => client.time()).to.throw('Rest module not provided');
+      });
+    });
   });
 });
